hw9-3.c

+++ hw9-3.c
@@ -1,21 +1,21 @@
 #include<stdio.h>
 int main()
 {
-	int a[2][3], c[2][2];
-	int  b[3][2];
-	int i, j;
+	double a[2][3], c[2][2];
+	double  b[3][2];
+	double i, j;
 	for(i = 0; i < 2; i++)
 		{
 			for(j=0; j<3; j++)
 			{
-				scanf("%d", &a[i][j]);
+				scanf("%lf", &a[i][j]);
 			}
 		}
 	for(i = 0; i < 3; i++)
 		{
 			for(j=0; j<2; j++)
 			{
-				scanf("%d", &b[i][j]);
+				scanf("%lf", &b[i][j]);
 			}
 		}
 	for(i = 0; i < 2; i++)
@@ -32,13 +32,14 @@ int main()
 	{
 		for(j=0; j<2; j++)
 		{
-			printf("%d ", c[i][j]);
+			printf("%.1f ", c[i][j]);
 		}
 	printf("\n");
 	}
 
 return 0;
 }
+
